requirements.txt
Flask>=2.1,<3.0
Werkzeug>=2.2,<3.0
Flask-PyMongo==2.3.0
pytest==7.4.0
python-dateutil>=2.8.1
python-dotenv